requirements.txt
packaging==23.1
Pillow==10.0.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pytesseract==0.3.10
